move setting of gitPrivateKey to init
