+++ lib/workers/repository/index.js
@@ -1,7 +1,6 @@
 const fs = require('fs-extra');
 const os = require('os');
 const path = require('path');
-const Git = require('simple-git/promise');
 
 const { initRepo } = require('./init');
 const { ensureOnboardingPr } = require('./onboarding/pr');
@@ -14,18 +13,6 @@ const { ensureMasterIssue } = require('./master-issue');
 module.exports = {
   renovateRepository,
 };
-
-async function setGitPrivateKey(config) {
-  if (config.gitPrivateKey && config.localDir) {
-    try {
-      const cwd = config.localDir;
-      const git = Git(cwd).silent(true);
-      await git.raw(['config', 'user.signingkey', config.gitPrivateKey]);
-    } catch (err) {
-      logger.warn({ err }, 'Error in setting git private key to author');
-    }
-  }
-}
 
 // istanbul ignore next
 async function renovateRepository(repoConfig) {
@@ -44,7 +31,6 @@ async function renovateRepository(repoConfig) {
       config.localDir || path.join(tmpDir, config.platform, config.repository);
     await fs.ensureDir(config.localDir);
     logger.debug('Using localDir: ' + config.localDir);
-    await setGitPrivateKey(config);
     config = await initRepo(config);
     const { res, branches, branchList, packageFiles } = await processRepo(
       config
