+++ lib/workers/repository/init/apis.js
@@ -1,3 +1,4 @@
+const Git = require('simple-git/promise');
 const { initPlatform } = require('../../../platform');
 const npmApi = require('../../../datasource/npm');
 
@@ -5,6 +6,18 @@ function assignPlatform(config) {
   logger.trace('assignPlatform');
   initPlatform(config.platform);
   return config;
+}
+
+async function setGitPrivateKey(config) {
+  if (config.gitPrivateKey && config.localDir) {
+    try {
+      const cwd = config.localDir;
+      const git = Git(cwd).silent(true);
+      await git.raw(['config', 'user.signingkey', config.gitPrivateKey]);
+    } catch (err) {
+      logger.warn({ err }, 'Error in setting git private key to author');
+    }
+  }
 }
 
 async function getPlatformConfig(config) {
@@ -17,6 +30,7 @@ async function getPlatformConfig(config) {
 
 async function initApis(input) {
   let config = { ...input };
+  await setGitPrivateKey(config);
   config = await assignPlatform(config);
   config = await getPlatformConfig(config);
   npmApi.resetMemCache();
